중간커밋. 장염이 이제서야 다 나은 거 같아요... 힘들었어요...

+++ src/main/java/portfolio/project/hashtagqna/service/QuestionService.java
@@ -8,10 +8,7 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import portfolio.project.hashtagqna.dto.QuestionDto;
 import portfolio.project.hashtagqna.dto.QuestionListDto;
-import portfolio.project.hashtagqna.entity.Hashtag;
-import portfolio.project.hashtagqna.entity.Member;
-import portfolio.project.hashtagqna.entity.Question;
-import portfolio.project.hashtagqna.entity.QuestionHashtag;
+import portfolio.project.hashtagqna.entity.*;
 import portfolio.project.hashtagqna.exception.AuthorizationExeption;
 import portfolio.project.hashtagqna.repository.HashtagRepository;
 import portfolio.project.hashtagqna.repository.QuestionHashtagRepository;
@@ -93,6 +90,9 @@ public class QuestionService {
 
     @Transactional
     public long updateQuestion(Question oldQuestion, Question editedQuestion, Member questionWriter) {
+        if (oldQuestion.getQuestionStatus() == QuestionStatus.CLOSED) {
+            throw new AuthorizationExeption("닫힌 질문은 수정이나 삭제할 수 없습니다.");
+        }
         if (oldQuestion.getMember() != questionWriter) {
             throw new AuthorizationExeption("질문 작성자만이 질문을 수정할 수 있습니다.");
         }
@@ -101,9 +101,20 @@ public class QuestionService {
 
     @Transactional
     public long removeQuestion(Question question, Member questionWriter) {
+        if (question.getQuestionStatus() == QuestionStatus.CLOSED) {
+            throw new AuthorizationExeption("닫힌 질문은 수정이나 삭제할 수 없습니다.");
+        }
         if (question.getMember() != questionWriter) {
             throw new AuthorizationExeption("질문 작성자만이 질문을 삭제할 수 있습니다.");
         }
+
+        for (int i = 0; i < question.getQuestionHashtags().size(); i++) {
+            questionWriter.decreaseHashTagCount();
+        }
+        for (Answer answer : question.getAnswers()) {
+            answer.getMember().decreaseAnswerCount();
+        }
+        questionWriter.decreaseQuestionCount();
         return questionRepository.removeQuestion(question);
     }
